aladdinsun/AIdatabase/requirements.txt
Flask==3.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
pytesseract==0.3.10
openpyxl==3.1.2
scikit-learn==1.3.2
transformers==4.35.2
easyocr==1.7.1
matplotlib==3.8.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
